store DB_PATH in static
Store the computed db_path in a LazyLock.
Use Cow<'static, Path> to prevent memory leaks of the previous Box::leak.
Fail when the passed path is invalid instead of silently ignoring it.

Note:
sqlx::MigrateDatabase takes &str as parameter, so the Path has to be converted when needed.
Even taking impl AsRef<str> would make it nicer, but that's just how it is.